pyproject.toml
[project]
name = "cloth-fusion"
version = "0.1.0"
description = "虚拟试衣间 Demo — Virtual Try-On Showcase"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "rembg[cpu]>=2.0.57",
    "pillow>=10.4.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
